docker: clarify usage changes re: alpine
Unfortunately, even the most recent version of Alpine (3.8) using the
available ChromeDriver and Chromium packages run into errors trying to
use Capybara and Selenium-webdriver for JS tests.

So we're going to change to doing local development with only postgres
in a container, assuming the developer will run the Puma app locally and
have chromedriver and chrome(ium) installed locally.