docker: switch to debian-based container for dev
The Alpine image is now still around as Dockerfile.prod in case we
decide to use that for future Kubernetes deployment.

In the meantime, we're using the standard ruby debian-based container
now which, while much larger in footprint, allows us to fairly easily
install current versions of nodejs, chrome, and chromedriver.

Since circleci uses a different setup which works just fine with the
normal :selenium_chrome_headless drive, we use that when in the CI
context.